+++ include/databento/dbn_decoder.hpp
@@ -25,15 +25,15 @@ class DbnDecoder {
              VersionUpgradePolicy upgrade_policy);
 
   static std::pair<std::uint8_t, std::size_t> DecodeMetadataVersionAndSize(
-      const std::uint8_t* buffer, std::size_t size);
+      const std::byte* buffer, std::size_t size);
   static Metadata DecodeMetadataFields(std::uint8_t version,
-                                       const std::vector<std::uint8_t>& buffer);
+                                       const std::vector<std::byte>& buffer);
   // Decodes a record possibly applying upgrading the data according to the
   // given version and upgrade policy. If an upgrade is applied,
   // compat_buffer is modified.
   static Record DecodeRecordCompat(
       std::uint8_t version, VersionUpgradePolicy upgrade_policy, bool ts_out,
-      std::array<std::uint8_t, kMaxRecordLen>* compat_buffer, Record rec);
+      std::array<std::byte, kMaxRecordLen>* compat_buffer, Record rec);
 
   // Should be called exactly once.
   Metadata DecodeMetadata();
@@ -44,19 +44,19 @@ class DbnDecoder {
  private:
   static std::string DecodeSymbol(
       std::size_t symbol_cstr_len,
-      std::vector<std::uint8_t>::const_iterator& buffer_it);
+      std::vector<std::byte>::const_iterator& buffer_it);
   static std::vector<std::string> DecodeRepeatedSymbol(
       std::size_t symbol_cstr_len,
-      std::vector<std::uint8_t>::const_iterator& buffer_it,
-      std::vector<std::uint8_t>::const_iterator buffer_end_it);
+      std::vector<std::byte>::const_iterator& buffer_it,
+      std::vector<std::byte>::const_iterator buffer_end_it);
   static std::vector<SymbolMapping> DecodeSymbolMappings(
       std::size_t symbol_cstr_len,
-      std::vector<std::uint8_t>::const_iterator& buffer_it,
-      std::vector<std::uint8_t>::const_iterator buffer_end_it);
+      std::vector<std::byte>::const_iterator& buffer_it,
+      std::vector<std::byte>::const_iterator buffer_end_it);
   static SymbolMapping DecodeSymbolMapping(
       std::size_t symbol_cstr_len,
-      std::vector<std::uint8_t>::const_iterator& buffer_it,
-      std::vector<std::uint8_t>::const_iterator buffer_end_it);
+      std::vector<std::byte>::const_iterator& buffer_it,
+      std::vector<std::byte>::const_iterator buffer_end_it);
   bool DetectCompression();
   std::size_t FillBuffer();
   std::size_t GetReadBufferSize() const;
@@ -67,11 +67,10 @@ class DbnDecoder {
   VersionUpgradePolicy upgrade_policy_;
   bool ts_out_{};
   std::unique_ptr<IReadable> input_;
-  std::vector<std::uint8_t> read_buffer_;
+  std::vector<std::byte> read_buffer_;
   std::size_t buffer_idx_{};
   // Must be 8-byte aligned for records
-  alignas(
-      RecordHeader) std::array<std::uint8_t, kMaxRecordLen> compat_buffer_{};
+  alignas(RecordHeader) std::array<std::byte, kMaxRecordLen> compat_buffer_{};
   Record current_record_{nullptr};
 };
 }  // namespace databento