+++ include/databento/iwritable.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <cstddef>  // size_t
-#include <cstdint>  // uint8_t
+#include <cstddef>  // byte, size_t
 
 namespace databento {
 // An abstract class for writable objects to allow for runtime polymorphism
@@ -10,6 +9,6 @@ class IWritable {
  public:
   virtual ~IWritable() = default;
 
-  virtual void WriteAll(const std::uint8_t* buffer, std::size_t length) = 0;
+  virtual void WriteAll(const std::byte* buffer, std::size_t length) = 0;
 };
 }  // namespace databento