+++ src/detail/json_helpers.cpp
@@ -2,6 +2,7 @@
 
 #include <numeric>  // accumulate
 #include <sstream>  // istringstream
+#include <string>
 
 namespace databento::detail {
 void SetIfNotEmpty(httplib::Params* params, const std::string& key,
@@ -25,9 +26,9 @@ void SetIfNotEmpty(httplib::Params* params, const std::string& key,
   }
 }
 
-const nlohmann::json& CheckedAt(const std::string& endpoint,
+const nlohmann::json& CheckedAt(std::string_view endpoint,
                                 const nlohmann::json& json,
-                                const std::string& key) {
+                                std::string_view key) {
   if (json.contains(key)) {
     return json.at(key);
   }
@@ -35,88 +36,91 @@ const nlohmann::json& CheckedAt(const std::string& endpoint,
 }
 
 template <>
-bool ParseAt(const std::string& endpoint, const nlohmann::json& json,
-             const std::string& key) {
+bool ParseAt(std::string_view endpoint, const nlohmann::json& json,
+             std::string_view key) {
   const auto& val_json = CheckedAt(endpoint, json, key);
   if (!val_json.is_boolean()) {
-    throw JsonResponseError::TypeMismatch(endpoint, key + " bool", val_json);
+    throw JsonResponseError::TypeMismatch(endpoint, std::string{key} + " bool",
+                                          val_json);
   }
   return val_json;
 }
 
 template <>
-std::string ParseAt(const std::string& endpoint, const nlohmann::json& json,
-                    const std::string& key) {
+std::string ParseAt(std::string_view endpoint, const nlohmann::json& json,
+                    std::string_view key) {
   const auto& val_json = CheckedAt(endpoint, json, key);
   if (val_json.is_null()) {
     return {};
   }
   if (!val_json.is_string()) {
-    throw JsonResponseError::TypeMismatch(endpoint, key + " string", val_json);
+    throw JsonResponseError::TypeMismatch(
+        endpoint, std::string{key} + " string", val_json);
   }
   return val_json;
 }
 
 template <>
-std::uint64_t ParseAt(const std::string& endpoint, const nlohmann::json& json,
-                      const std::string& key) {
+std::uint64_t ParseAt(std::string_view endpoint, const nlohmann::json& json,
+                      std::string_view key) {
   const auto& val_json = CheckedAt(endpoint, json, key);
   if (val_json.is_null()) {
     return 0;
   }
   if (!val_json.is_number_unsigned()) {
-    throw JsonResponseError::TypeMismatch(endpoint, key + " unsigned number",
-                                          val_json);
+    throw JsonResponseError::TypeMismatch(
+        endpoint, std::string{key} + " unsigned number", val_json);
   }
   return val_json;
 }
 
 template <>
-std::uint16_t ParseAt(const std::string& endpoint, const nlohmann::json& json,
-                      const std::string& key) {
+std::uint16_t ParseAt(std::string_view endpoint, const nlohmann::json& json,
+                      std::string_view key) {
   const auto& val_json = CheckedAt(endpoint, json, key);
   if (val_json.is_null()) {
     return 0;
   }
   if (!val_json.is_number_unsigned()) {
-    throw JsonResponseError::TypeMismatch(endpoint, key + " unsigned number",
-                                          val_json);
+    throw JsonResponseError::TypeMismatch(
+        endpoint, std::string{key} + " unsigned number", val_json);
   }
   return val_json;
 }
 
 template <>
-double ParseAt(const std::string& endpoint, const nlohmann::json& json,
-               const std::string& key) {
+double ParseAt(std::string_view endpoint, const nlohmann::json& json,
+               std::string_view key) {
   const auto& val_json = CheckedAt(endpoint, json, key);
   if (val_json.is_null()) {
     return 0;
   }
   if (!val_json.is_number()) {
-    throw JsonResponseError::TypeMismatch(endpoint, key + " number", val_json);
+    throw JsonResponseError::TypeMismatch(
+        endpoint, std::string{key} + " number", val_json);
   }
   return val_json;
 }
 
 template <>
-std::vector<std::string> ParseAt(const std::string& endpoint,
+std::vector<std::string> ParseAt(std::string_view endpoint,
                                  const nlohmann::json& json,
-                                 const std::string& key) {
+                                 std::string_view key) {
   const auto& symbols_json = CheckedAt(endpoint, json, key);
   // if there's only one symbol, it returns a string not an array
   if (symbols_json.is_string()) {
     return {symbols_json};
   }
   if (!symbols_json.is_array()) {
-    throw JsonResponseError::TypeMismatch(endpoint, key + " array", json);
+    throw JsonResponseError::TypeMismatch(endpoint, std::string{key} + " array",
+                                          json);
   }
   return {symbols_json.begin(), symbols_json.end()};
 }
 
 template <>
-date::year_month_day ParseAt(const std::string& endpoint,
-                             const nlohmann::json& json,
-                             const std::string& key) {
+date::year_month_day ParseAt(std::string_view endpoint,
+                             const nlohmann::json& json, std::string_view key) {
   std::string raw_start = detail::CheckedAt(endpoint, json, key);
   std::istringstream start_stream{raw_start};
   date::year_month_day start;