+++ tests/include/mock/mock_io.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <cstddef>
-#include <cstdint>
 #include <vector>
 
 #include "databento/ireadable.hpp"
@@ -10,16 +9,16 @@
 namespace databento::tests::mock {
 class MockIo : public databento::IWritable, public databento::IReadable {
  public:
-  void WriteAll(const std::uint8_t* buffer, std::size_t length);
+  void WriteAll(const std::byte* buffer, std::size_t length);
 
-  void ReadExact(std::uint8_t* buffer, std::size_t length);
+  void ReadExact(std::byte* buffer, std::size_t length);
 
-  std::size_t ReadSome(std::uint8_t* buffer, std::size_t max_length);
+  std::size_t ReadSome(std::byte* buffer, std::size_t max_length);
 
-  const std::vector<std::uint8_t>& GetContents() const { return contents_; }
+  const std::vector<std::byte>& GetContents() const { return contents_; }
 
  private:
-  std::vector<std::uint8_t> contents_;
+  std::vector<std::byte> contents_;
   std::ptrdiff_t read_idx_{0};
 };
 }  // namespace databento::tests::mock